requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
numpy==1.21.2
opencv-python==4.5.3.56
passlib==1.7.4
Pillow==8.3.1
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.2
Werkzeug==2.0.1
